requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.1
python-socketio==5.0.4
python-dotenv==0.15.0